Add Game constructor and random phrase to Phrase.
This commit creates the Game class constructor, as
well as implements a way to randomly generate a
phrase in the Phrase class.  Relies on an ArrayRepo
argument required to be passed into the call to instantiate
the Phrase object (this seems cleaner that housing
the entire data store in the Phrase class.